Move integration tests to a dedicated folder
